+++ example/bin/number_example.dart
@@ -1,0 +1,10 @@
+import 'package:siberia/siberia.dart';
+
+void main(List<String> arguments) {
+  int? exampleInt;
+  double? exampleDouble;
+
+  //orZero example
+  print(exampleInt.orZero);
+  print(exampleDouble.orZero);
+}