+++ lib/src/primitives/number/double_extensions.dart
@@ -1,0 +1,4 @@
+extension DoubleExtension on double? {
+  /// returns 0 if input value is null
+  double get orZero => this ?? 0.0;
+}