+++ lib/src/primitives/number/number.dart
@@ -1,0 +1,4 @@
+library number;
+
+export 'double_extensions.dart';
+export 'int_extensions.dart';