+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.1.0 - Saturday, Sep. 18, 2021
+- fix: normalized library export files
+- feat: added github actions pull request validation flow
+
+**BREAKING CHANGES**
+- feat: Changed some extension methods into getters
+  - orEmpty() => orEmpty;
+  - orZero() => orZero;
+  - toIntOrNull() => tryToInt();
+  - toDoubleOrNull() => tryToDouble();
+
 ## 1.0.1
 
 Rename files. It's a breaking change, need to update the imports.